feat: Create HTML report
make report nicer

Issue #7
