+++ pmlib/report/__init__.py
@@ -22,133 +22,17 @@ import pmlib
 
 from yattag import Doc, indent
 from pmlib.item import Item
+from pmlib.report.css import report_style
+from pmlib.report.js import report_script
 from pmlib.types import Entry
 
 
 __all__ = [
+    "css",
+    "js",
+
     "Report"
 ]
-
-__css__ = """
-body {
-    font-family: Arial, Helvetica, sans-serif;
-    font-size: 14px;
-}
-
-table {
-    border-spacing: 5px;
-}
-
-ul, #myUL {
-    list-style-type: none;
-}
-
-#myUL {
-    margin: 0;
-    padding: 0;
-}
-
-.caret {
-    cursor: pointer;
-    -webkit-user-select: none; /* Safari 3.1+ */
-    -moz-user-select: none; /* Firefox 2+ */
-    -ms-user-select: none; /* IE 10+ */
-    user-select: none;
-}
-
-.caret::before {
-    content: '\25B6';
-    color: black;
-    display: inline-block;
-    margin-right: 6px;
-}
-
-.caret-down::before {
-    -ms-transform: rotate(90deg); /* IE 9 */
-    -webkit-transform: rotate(90deg); /* Safari */'
-    transform: rotate(90deg);  
-}
-
-.nested {
-    display: none;
-}
-
-.activetray {
-    display: block;
-}
-
-.folder {
-    font-family: Arial, Helvetica, sans-serif;
-    padding: 6px;
-    color: blue;
-    cursor: pointer;
-    width: 20%;
-    border: none;
-    text-align: left;
-    outline: none;
-    font-size: 14px;
-}
-
-.tray {
-    font-family: Arial, Helvetica, sans-serif;
-    padding: 6px;
-    cursor: pointer;
-    width: 20%;
-    border: none;
-    text-align: left;
-    outline: none;
-    font-size: 14px;
-}
-
-.active, .folder:hover {
-    font-family: Arial, Helvetica, sans-serif;
-    background-color: #00FFFF;
-    padding: 12px;
-    font-weight: bold;
-}
-
-.content {
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-    padding-top: 20px; 
-    padding-left: 20px; 
-    padding-bottom: 20px; 
-    display: none;
-    overflow: hidden;
-    width: 80%;
-    background-color: #F0FFFF;
-    margin-bottom: 30px;
-}
-
-.firstcolumn {
-    width: 30%;
-    font-weight: bold;
-}
-"""
-
-__script__ = """
-var toggler = document.getElementsByClassName("caret");
-var coll = document.getElementsByClassName("collapsible");
-var i;
-
-for (i = 0; i < toggler.length; i++) {
-  toggler[i].addEventListener("click", function() {
-    this.parentElement.querySelector(".nested").classList.toggle("activetray");
-    this.classList.toggle("caret-down");
-  });
-}
-
-for (i = 0; i < coll.length; i++) {
-  coll[i].addEventListener("click", function() {
-    this.classList.toggle("active");
-    var content = this.nextElementSibling;
-    if (content.style.display === "block") {
-      content.style.display = "none";
-    } else {
-      content.style.display = "block";
-    }
-  });
-}
-"""
 
 
 def _sort_name(item: Item):
@@ -172,7 +56,7 @@ class Report(object):
 
             doc.stag('meta', name='viewport', content='width=device-width, initial-scale=1')
             with tag("style"):
-                text(__css__)
+                text(report_style)
 
         return
 
@@ -189,9 +73,10 @@ class Report(object):
                 with tag("tr"):
                     line("td", "Success", klass="firstcolumn")
                     line("td", "{0:d}".format(item.report.success))
-                with tag("tr"):
-                    line("td", "failure", klass="firstcolumn")
-                    line("td", "{0:d}".format(item.report.failure))
+                if item.report.failure != 0:
+                    with tag("tr"):
+                        line("td", "Failure", klass="firstfailure")
+                        line("td", "{0:d}".format(item.report.failure))
         return
 
     def _create_item(self, li, item: Item):
@@ -205,14 +90,14 @@ class Report(object):
                 for _item in sorted(item.children, key=_sort_name):
                     if _item.type is Entry.folder:
                         with tag("li"):
-                            with tag("button", type="button", klass="collapsible"):
-                                text(_item.name)
+                            with tag("button", type="button", klass="folder"):
+                                doc.asis("&#9670; {0:s}".format(_item.name))
                             div = tag("div", klass="content")
                             self._create_report(div, _item)
 
                 for _item in sorted(item.children, key=_sort_name):
                     if _item.type is Entry.tray:
-                        _li = tag("li")
+                        _li = tag("li", klass="tray")
                         self._create_item(_li, _item)
         return
 
@@ -228,7 +113,7 @@ class Report(object):
                 self._create_item(root, self.root)
 
             with tag("script"):
-                doc.asis(__script__)
+                doc.asis(report_script)
 
         return
